requirements.txt
Flask==3.0.2
mysql-connector-python==8.2.0
cryptography==41.0.7
python-dotenv==1.0.1
Flask-Session==0.5.0
